+++ spec/unit/query/preload/association-on-other-side-of-polymorphic-belongs-to.spec.ts
@@ -2,15 +2,22 @@ import { sql } from 'kysely'
 import db from '../../../../src/db/index.js'
 import Chore from '../../../../test-app/app/models/Polymorphic/Chore.js'
 import ChoreCleaningSupply from '../../../../test-app/app/models/Polymorphic/ChoreCleaningSupply.js'
+import ChoreImage from '../../../../test-app/app/models/Polymorphic/ChoreImage.js'
 import CleaningSupply from '../../../../test-app/app/models/Polymorphic/CleaningSupply.js'
+import Image from '../../../../test-app/app/models/Polymorphic/Image.js'
 import PolymorphicTask from '../../../../test-app/app/models/Polymorphic/Task.js'
+import TaskableImage from '../../../../test-app/app/models/Polymorphic/TaskableImage.js'
 import PolymorphicUser from '../../../../test-app/app/models/Polymorphic/User.js'
 import Workout from '../../../../test-app/app/models/Polymorphic/Workout.js'
+import WorkoutImage from '../../../../test-app/app/models/Polymorphic/WorkoutImage.js'
+import WorkoutType from '../../../../test-app/app/models/Polymorphic/WorkoutType.js'
 
 describe('preloading associations on the other side of a polymorphic belongs-to', () => {
   let chore: Chore
   let workout: Workout
+
   let cleaningSupply: CleaningSupply
+  let workoutType: WorkoutType
 
   let choreTask: PolymorphicTask
   let workoutTask: PolymorphicTask
@@ -26,9 +33,11 @@ describe('preloading associations on the other side of a polymorphic belongs-to'
     ])
 
     const user = await PolymorphicUser.create()
-    chore = await Chore.create()
-    workout = await Workout.create()
     cleaningSupply = await CleaningSupply.create()
+    workoutType = await WorkoutType.create({ workoutType: 'cycling' })
+
+    chore = await Chore.create()
+    workout = await Workout.create({ workoutType })
 
     await ChoreCleaningSupply.create({ chore, cleaningSupply })
 
@@ -40,7 +49,15 @@ describe('preloading associations on the other side of a polymorphic belongs-to'
     'allows specifying of associations that only exist on some of the associations on the other side ' +
       'of the polymorphic belongs-to',
     async () => {
-      const user = await PolymorphicUser.preload('tasks', 'taskable', ['cleaningSupplies']).firstOrFail()
+      const user = await PolymorphicUser.preload('tasks', 'taskable', [
+        'cleaningSupplies',
+        /**
+         * TODO: figure out why types disallow loading associations on the other side of a
+         * polymorphic belongs-to association other than those associated with the first
+         * of the polymorphic classes (first as ordered by table name)
+         */
+        'workoutType' as any,
+      ]).firstOrFail()
 
       const loadedChoreTask = user.tasks.find(
         (task): task is PolymorphicTask & { taskable: Chore } => task.taskableType === 'Chore'
@@ -54,7 +71,9 @@ describe('preloading associations on the other side of a polymorphic belongs-to'
 
       expect(loadedChoreTask.taskable).toMatchDreamModel(chore)
       expect(loadedWorkoutTask.taskable).toMatchDreamModel(workout)
+
       expect(loadedChoreTask.taskable.cleaningSupplies).toMatchDreamModels([cleaningSupply])
+      expect(loadedWorkoutTask.taskable.workoutType).toMatchDreamModel(workoutType)
     }
   )
 
@@ -70,5 +89,47 @@ describe('preloading associations on the other side of a polymorphic belongs-to'
 
     expect(loadedChoreTask.taskable.cleaningSupplies).toMatchDreamModels([cleaningSupply])
     expect(loadedWorkoutTask.taskable.cleaningSupplies).toBeUndefined()
+  })
+
+  context('has many on the other side of a polymorphic belongs-to', () => {
+    it('loads through a non-polymorphic association', async () => {
+      const choreImage = await Image.create({ url: 'https://images.com/chore.png' })
+      const workoutImage = await Image.create({ url: 'https://images.com/workout.png' })
+      await ChoreImage.create({ image: choreImage, chore })
+      await WorkoutImage.create({ image: workoutImage, workout })
+
+      const user = await PolymorphicUser.preload('tasks', 'taskable', ['images']).firstOrFail()
+
+      const loadedChoreTask = user.tasks.find(
+        (task): task is PolymorphicTask & { taskable: Chore } => task.taskableType === 'Chore'
+      )!
+      const loadedWorkoutTask = user.tasks.find(
+        (task): task is PolymorphicTask & { taskable: Chore } => task.taskableType === 'Workout'
+      )!
+
+      expect(loadedChoreTask.taskable.images).toMatchDreamModels([choreImage])
+      expect(loadedWorkoutTask.taskable.images).toMatchDreamModels([workoutImage])
+    })
+
+    it('loads through a polymorphic association', async () => {
+      const choreImage = await Image.create({ url: 'https://images.com/chore.png' })
+      const workoutImage = await Image.create({ url: 'https://images.com/workout.png' })
+      await TaskableImage.create({ image: choreImage, taskable: chore })
+      await TaskableImage.create({ image: workoutImage, taskable: workout })
+
+      const user = await PolymorphicUser.preload('tasks', 'taskable', [
+        'imagesThroughTaskableImages',
+      ]).firstOrFail()
+
+      const loadedChoreTask = user.tasks.find(
+        (task): task is PolymorphicTask & { taskable: Chore } => task.taskableType === 'Chore'
+      )!
+      const loadedWorkoutTask = user.tasks.find(
+        (task): task is PolymorphicTask & { taskable: Chore } => task.taskableType === 'Workout'
+      )!
+
+      expect(loadedChoreTask.taskable.imagesThroughTaskableImages).toMatchDreamModels([choreImage])
+      expect(loadedWorkoutTask.taskable.imagesThroughTaskableImages).toMatchDreamModels([workoutImage])
+    })
   })
 })