+++ src/dream-app/cache.ts
@@ -1,3 +1,4 @@
+import EnvInternal from '../helpers/EnvInternal.js'
 import DreamApp from './index.js'
 
 let _dreamApp: DreamApp | undefined = undefined
@@ -7,9 +8,20 @@ export function cacheDreamApp(dreamconf: DreamApp) {
 }
 
 export function getCachedDreamAppOrFail() {
-  if (!_dreamApp)
-    throw new Error(
-      'Must call `initializeDreamApp` or `initializePsychicApp` before loading cached DreamApp.getOrFail()'
-    )
+  if (!_dreamApp) {
+    const baseErrorMessage =
+      'Must call `initializeDreamApp` or `initializePsychicApp` before loading cached DreamApp.getOrFail().'
+
+    if (EnvInternal.isTest) {
+      throw new Error(
+        `${baseErrorMessage}
+
+Check for specs running directly in a \`describe\` or \`context\` block rather than nested within an \`it\` block.`
+      )
+    } else {
+      throw new Error(baseErrorMessage)
+    }
+  }
+
   return _dreamApp
 }