+++ test-app/app/models/Polymorphic/Workout.ts
@@ -2,7 +2,10 @@ import { Decorators } from '../../../../src/index.js'
 import { DreamColumn } from '../../../../src/types/dream.js'
 import { WorkoutSerializer } from '../../serializers/Polymorphic/WorkoutSerializer.js'
 import ApplicationModel from '../ApplicationModel.js'
+import Image from './Image.js'
 import PolymorphicLocalizedText from './LocalizedText.js'
+import TaskableImage from './TaskableImage.js'
+import WorkoutType from './WorkoutType.js'
 
 const deco = new Decorators<typeof Workout>()
 
@@ -22,9 +25,28 @@ export default class Workout extends ApplicationModel {
   public createdAt: DreamColumn<Workout, 'createdAt'>
   public updatedAt: DreamColumn<Workout, 'updatedAt'>
 
+  @deco.BelongsTo('Polymorphic/WorkoutType')
+  public workoutType: WorkoutType
+  public polymorphicWorkoutTypeId: DreamColumn<Workout, 'polymorphicWorkoutTypesId'>
+
   @deco.HasMany('Polymorphic/LocalizedText', {
     polymorphic: true,
     foreignKey: 'localizableId',
   })
   public localizedTexts: PolymorphicLocalizedText[]
+
+  @deco.HasMany('Polymorphic/TaskableImage', {
+    polymorphic: true,
+    foreignKey: 'taskableId',
+  })
+  public taskableImages: TaskableImage[]
+
+  @deco.HasMany('Polymorphic/Image', { through: 'taskableImages', source: 'image' })
+  public imagesThroughTaskableImages: Image[]
+
+  @deco.HasMany('Polymorphic/WorkoutImage')
+  public workoutImages: TaskableImage[]
+
+  @deco.HasMany('Polymorphic/Image', { through: 'workoutImages' })
+  public images: Image[]
 }