+++ test-app/db/migrations/1751977371079-create-polymorphic-image.ts
@@ -2,17 +2,14 @@ import { Kysely } from 'kysely'
 
 export async function up(db: Kysely<any>): Promise<void> {
   await db.schema
-    .createTable('polymorphic_workouts')
+    .createTable('polymorphic_images')
     .addColumn('id', 'bigserial', col => col.primaryKey())
-    .addColumn('polymorphic_workout_type_id', 'bigint', col =>
-      col.references('polymorphic_workout_types.id').onDelete('restrict').notNull()
-    )
-    .addColumn('name', 'varchar(255)')
+    .addColumn('url', 'text', col => col.notNull())
     .addColumn('created_at', 'timestamp', col => col.notNull())
     .addColumn('updated_at', 'timestamp', col => col.notNull())
     .execute()
 }
 
 export async function down(db: Kysely<any>): Promise<void> {
-  await db.schema.dropTable('polymorphic_workouts').execute()
+  await db.schema.dropTable('polymorphic_images').execute()
 }